+++ src/Exceptions/InvalidPayloadException.php
@@ -7,4 +7,4 @@ namespace Paytabs\Laravel\Exceptions;
 /**
  * Thrown when a callback payload cannot be mapped or is not of the expected type.
  */
-class InvalidPayloadException extends IpnProcessingException {}
+class InvalidPayloadException extends CallbackProcessingException {}